requirements.txt
Django==2.2
djangorestframework==3.9.2
numpy==1.16.2
pandas==0.24.2
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
xlrd==1.2.0
